refactored and documented BroadcastReceiver
